requirements.txt
treamlit
import
spacy
nltk
io
sys
streamlit_lottie
requests